Move BoundsStream to bounds mod
